Refactor matrix!, stack! macros into separate modules
